+++ src/components/Form/Title/styles.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const Title = styled.h1`
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 1rem;
+`;
