+++ src/interfaces/UserAPI.d.ts
@@ -1,17 +1,9 @@
-import { CommentAPI } from './CommentAPI.d';
+import { CommentAPI } from './CommentAPI';
 import { PostAPI } from './PostAPI';
 
-export interface NotificationAPI {
-  seen: boolean;
+export interface UserInfo {
   _id: string;
-  author: User;
-  user: User | string;
-  post: Nullable<string>; // 포스트 id
-  follow: string | undefined; // 사용자 id
-  comment: string | undefined;
-  message: string | undefined; // 메시지 id
-  createdAt: string;
-  updatedAt: string;
+  points: number;
 }
 
 export interface followingType {
